requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-filter
mysqlclient
django-cors-headers
